added sparky_types.h

+++ Sparky/src/graphics/font.cpp
@@ -4,8 +4,8 @@ namespace sparky
 {
 	namespace graphics
 	{
-		Font::Font(std::string name, std::string filename, unsigned int size)
-			: m_Name(name), m_Filename(filename), m_Size(size), m_Scale(maths::vec2(1, 1))
+		Font::Font(std::string name, std::string filename, float size)
+			: m_Name(name), m_Filename(filename), m_Size(size), m_Scale(maths::vec2(1.0f, 1.0f))
 		{
 			m_FTAtlas = ftgl::texture_atlas_new(512, 512, 2);
 			m_FTFont = ftgl::texture_font_new_from_file(m_FTAtlas, size, filename.c_str());
