+++ Sparky/src/graphics/font_manager.h
@@ -18,7 +18,7 @@ namespace sparky
 			static void add(Font *font);
 			static Font *get();
 			static Font *get(const std::string &name);
-			static Font *get(const std::string &name, unsigned int size);
+			static Font *get(const std::string &name, float size);
 			static void clean();
 		};
 	}
