+++ Sparky/src/graphics/label.cpp
@@ -4,21 +4,21 @@ namespace sparky
 {
 	namespace graphics
 	{
-		Label::Label(std::string text, float x, float y, unsigned int color)
+		Label::Label(std::string text, float x, float y, uint color)
 			: Renderable2D(), text(text), x(x), y(y), position(m_Position), m_Font(FontManager::get("SourceSansPro"))
 		{
 			m_Position = maths::vec3(x, y, 0.0f);
 			m_Color = color;
 		};
 
-		Label::Label(std::string text, float x, float y, Font *font, unsigned int color)
+		Label::Label(std::string text, float x, float y, Font *font, uint color)
 			: Renderable2D(), text(text), x(x), y(y), position(m_Position), m_Font(font)
 		{
 			m_Position = maths::vec3(x, y, 0.0f);
 			m_Color = color;
 		};
 
-		Label::Label(std::string text, float x, float y, const std::string &fontname, unsigned int color)
+		Label::Label(std::string text, float x, float y, const std::string &fontname, uint color)
 			: Renderable2D(), text(text), x(x), y(y), position(m_Position), m_Font(FontManager::get(fontname))
 		{
 			m_Position = maths::vec3(x, y, 0.0f);
@@ -27,7 +27,7 @@ namespace sparky
 			validatefont(fontname);
 		};
 
-		Label::Label(std::string text, float x, float y, const std::string &fontname, unsigned int size, unsigned int color)
+		Label::Label(std::string text, float x, float y, const std::string &fontname, float size, uint color)
 			: Renderable2D(), text(text), x(x), y(y), position(m_Position), m_Font(FontManager::get(fontname, size))
 		{
 			m_Position = maths::vec3(x, y, 0.0f);
@@ -36,7 +36,7 @@ namespace sparky
 			validatefont(fontname, size);
 		};
 
-		void Label::validatefont(const std::string &name, unsigned int size)
+		void Label::validatefont(const std::string &name, float size)
 		{
 			if (m_Font != nullptr)
 				return;
