+++ Sparky/src/graphics/renderable2d.h
@@ -17,7 +17,7 @@ namespace sparky
 			maths::vec3 vertex;
 			maths::vec2 uv; // texture coordinates
 			float tid;
-			unsigned int color;
+			uint color;
 		};
 
 		class Renderable2D
@@ -25,14 +25,14 @@ namespace sparky
 		protected:
 			maths::vec3 m_Position;
 			maths::vec2 m_Size;
-			unsigned int m_Color;
+			uint m_Color;
 			std::vector<maths::vec2> m_UV;
 			Texture *m_Texture;
 
 			Renderable2D();
 
 		public:
-			Renderable2D(maths::vec3 position, maths::vec2 size, unsigned int color);
+			Renderable2D(maths::vec3 position, maths::vec2 size, uint color);
 
 			virtual ~Renderable2D();
 
@@ -41,23 +41,23 @@ namespace sparky
 				renderer->submit(this);
 			};
 
-			void setColor(unsigned int color)
+			void setColor(uint color)
 			{
 				m_Color = color;
 			}
 			void setColor(const maths::vec4 &color)
 			{
-				int r = color.x * 255.0f;
-				int g = color.y * 255.0f;
-				int b = color.z * 255.0f;
-				int a = color.w * 255.0f;
+				uint r = (uint)(color.x * 255.0f);
+				uint g = (uint)(color.y * 255.0f);
+				uint b = (uint)(color.z * 255.0f);
+				uint a = (uint)(color.w * 255.0f);
 
 				m_Color = a << 24 | b << 16 | g << 8 | r;
 			}
 
 			inline const maths::vec3 &getPosition() const { return m_Position; };
 			inline const maths::vec2 &getSize() const { return m_Size; };
-			inline const unsigned int getColor() const { return m_Color; };
+			inline const uint getColor() const { return m_Color; };
 			inline const std::vector<maths::vec2> &getUV() const { return m_UV; };
 
 			inline const GLuint getTID() const { return m_Texture ? m_Texture->getID() : 0; };
